Refactored Deck into data access with and implemented an interface through which to access the api
