+++ src/main/java/data_access/DeckAPIInterface.java
@@ -1,0 +1,18 @@
+package data_access;
+
+import entity.Card;
+
+import java.util.List;
+
+public interface DeckAPIInterface {
+
+    class UnableToLoadDeck extends Exception{
+        //Exception thrown when API call fails
+        public UnableToLoadDeck(){
+            super("Unable to load deck");
+        }
+    }
+
+    //Draws n cards
+    public List<Card> drawCards(int n) throws UnableToLoadDeck;
+}
